LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/no-referrer-from-meta-url-censored-expected.txt



FAIL The url of a document with no-referrer referrer meta tag is censored in NavigationHistoryEntry assert_equals: expected 2 but got 0
